+++ tests/sample-tree/box/decoy.py
@@ -1,0 +1,1 @@
+# not imported actually!  this is not python 2 with implicit relative imports!